# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/participants/{id}": {
      "get": {
        "operationId": "AdminParticipantsController_fetchParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin / Participants"
        ]
      }
    },
    "/admin/participants": {
      "get": {
        "operationId": "AdminParticipantsController_listParticipants",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shift",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin / Participants"
        ]
      }
    },
    "/app/participants": {
      "post": {
        "operationId": "AppParticipantsController_createParticipant",
        "parameters": [
          {
            "name": "x-invitation-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreateParticipantInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppCreateParticipantCommandOutput"
                }
              }
            }
          }
        },
        "tags": [
          "App / Participants"
        ]
      }
    },
    "/app/participants/{id}": {
      "patch": {
        "operationId": "AppParticipantsController_updateParticipant",
        "parameters": [
          {
            "name": "x-invitation-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdateParticipantInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUpdateParticipantCommandOutput"
                }
              }
            }
          }
        },
        "tags": [
          "App / Participants"
        ]
      }
    },
    "/app/participants/by-invitation/{code}": {
      "get": {
        "operationId": "AppParticipantsController_fetchByInvitationCode",
        "parameters": [
          {
            "name": "x-invitation-code",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppFetchParticipantByInvitationCodeQueryOutput"
                }
              }
            }
          }
        },
        "tags": [
          "App / Participants"
        ]
      }
    },
...
  },
  "info": {
    "title": "Machine.off API",
    "description": "API for event invitation and RSVP management",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AppCreateParticipantInputDto": {
        "type": "object",
        "properties": {
          "invitationCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "shifts": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "lunch",
                "dinner"
              ]
            }
          },
          "acceptsImageUsage": {
            "type": "boolean"
          },
          "hasDietaryRestriction": {
            "type": "boolean"
          },
          "dietaryRestrictionDescription": {
            "type": "string"
          }
        },
        "required": [
          "invitationCode",
          "name",
          "email",
          "phone",
          "shifts",
          "acceptsImageUsage",
          "hasDietaryRestriction"
        ]
      },
      "AppCreateParticipantCommandOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt"
        ]
      },
      "AppUpdateParticipantInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "shifts": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "lunch",
                "dinner"
              ]
            }
          },
          "acceptsImageUsage": {
            "type": "boolean"
          },
          "hasDietaryRestriction": {
            "type": "boolean"
          },
          "dietaryRestrictionDescription": {
            "type": "string"
          }
        }
      },
      "AppUpdateParticipantCommandOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "updatedAt"
        ]
      },
      "AppParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invitationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "shifts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptsImageUsage": {
            "type": "boolean"
          },
          "hasDietaryRestriction": {
            "type": "boolean"
          },
          "dietaryRestrictionDescription": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "invitationId",
          "name",
          "email",
          "phone",
          "shifts",
          "acceptsImageUsage",
          "hasDietaryRestriction",
          "dietaryRestrictionDescription",
          "createdAt",
          "updatedAt"
        ]
      },
      "AppFetchParticipantByInvitationCodeQueryOutput": {
        "type": "object",
        "properties": {
          "participant": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AppParticipantDto"
              }
            ]
          }
        }
      },
...
    }
  }
}